Lots of Updates, Home Stretch
Lots of updates, added bunch of db functions for exporting. added that logic to the GUI too. Cleaned up a lot of code  here and there. Added the actual export to csv capability for four buttons. Parsed out a reporting table from the activities table. Probably need to just do all this for a "ProcessedStreams" table and then it will be strictly documentation and code clean up for the last two weeks.